requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
attrs==24.3.0
breadability==0.1.20
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
distro==1.9.0
Django==5.1.4
django-allauth==65.3.0
django-extensions==3.2.3
djangorestframework==3.15.2
docopt==0.6.2
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.12.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jwt==1.3.1
llvmlite==0.43.0
lxml==5.3.0
MarkupSafe==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
nltk==3.9.1
numba==0.60.0
numpy==2.0.2
openai-whisper==20240930
packaging==24.2
pillow==11.0.0
propcache==0.2.1
psycopg2-binary==2.9.10
pycountry==24.6.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.4.0
python-decouple==3.8
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.2
sumy==0.11.0
sympy==1.13.1
tiktoken==0.8.0
torch==2.5.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
whisper==1.1.10
yarl==1.18.3
yt-dlp==2024.12.13
